[DF-9566] Warning fix.
Former-commit-id: 612d3aa7ecb77d2e10818d98c158618ae19791c3

+++ Sources/Internal/FileSystem/StreamBuffer.cpp
@@ -91,7 +91,7 @@ uint32 StreamBuffer::ReadInternal(uint8* dataOut, uint32 len)
         return 0;
     }
 
-    uint32 dataSize = pages.front().size() - currentPageReadPos;
+    uint32 dataSize = static_cast<uint32>(pages.front().size()) - currentPageReadPos;
     uint32 sizeToRead = Min(dataSize, len);
     Memcpy(dataOut, pages.front().data() + currentPageReadPos, sizeToRead);
